Pad page with at least 0.5rem on all resolutions
Apply the padding of 0.5rem globally. It used to be applied only on small
(<=640px wide) screens. It mostly does not matter on wide screens -
the content of the page has a width limit and will be centered
when this limit is exceeded. But it used to look bad when the page
got zoomed in.
In any case, the new rules apply minimum viable padding uniformly
- problem solved.